chore: refactoring and cleanup
