+++ src/main.rs
@@ -4,6 +4,8 @@ use app::LogoMenu;
 mod app;
 mod config;
 mod core;
+mod logos;
+mod power;
 
 fn main() -> cosmic::iced::Result {
     cosmic::applet::run::<LogoMenu>(())
